+++ packages/components/organisms/f-cookie-banner/f-cookie-banner-static/src/App.vue
@@ -1,0 +1,14 @@
+<template>
+    <cookie-banner locale='en-NZ' />
+</template>
+
+<script>
+import CookieBanner from '@justeat/f-cookie-banner';
+import '@justeat/f-cookie-banner/dist/f-cookie-banner.css';
+
+export default {
+    components: {
+        CookieBanner
+    }
+};
+</script>