+++ packages/components/organisms/f-cookie-banner/f-cookie-banner-static/vue.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+    chainWebpack: config => {
+        config.module.rules.delete('eslint');
+    },
+    pluginOptions: {
+        prerenderSpa: {
+            registry: undefined,
+            renderRoutes: [
+                '/'
+            ],
+            useRenderEvent: true,
+            headless: true,
+            onlyProduction: true
+        }
+    }
+};