+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -13,6 +13,25 @@ v0.155.0
 - Added a max length of 50 on email.
 
 
+v0.152.0
+------------------------------
+*July 6, 2021*
+
+### Fixed
+- Returning guest user's mobile phone number not populating
+
+
+v0.151.0
+------------------------------
+*July 5, 2021*
+
+### Fixed
+- Get new `availableFulfilment` times when no available fulfilment times are available.
+
+### Changed
+- `f-form-field` version to `1.16.1`.
+
+
 v0.150.0
 ------------------------------
 *June 30, 2021*